switch home and project page
