+++ frontend/src/containers/Home.jsx
@@ -2,29 +2,21 @@ import React, { useState, useEffect } from "react"
 import axios from "axios"
 
 //components
-import {PageLayout} from '../components/PageLayout';
-// import Swipable from "../components/Swipable"
-// import Tasks from "../components/Tasks"
-// import MonthlyCalendar from "../components/Calendar"
-import SessionList from "../components/SessionList"
+import { PageLayout } from '../components/PageLayout';
+import Tasks from "../components/Tasks"
 
-const sessions = [
-  { id: 1, duration: '00:00:30', time_of_day: '2020-06-06', productivity: 'TRUE', site_id: 1 },
-  { id: 2, duration: '01:00:30', time_of_day: '2020-06-06', productivity: 'FALSE', site_id: 2 },
-  { id: 3, duration: '02:00:30', time_of_day: '2020-06-06', productivity: 'TRUE', site_id: 1 },
-  { id: 4, duration: '00:10:30', time_of_day: '2020-06-07', productivity: 'TRUE', site_id: 4 },
-  { id: 5, duration: '00:30:30', time_of_day: '2020-06-07', productivity: 'FALSE', site_id: 1 }
-]
 
 const HomePage = () => {
   const [state, setState] = useState({});
 
-  const fetchData = (
-    axios
-      .get(`localhost:8000/sessions/${user_id}`)
-      .then(data => setState(data))
-      .catch(error => console.log(error))
-  )
+  const fetchData = () => {
+    return (
+      axios
+        .get(`localhost:8000/projects/${user_id}`)
+        .then(data => setState(data))
+        .catch(error => console.log(error))
+    )
+  }
 
   useEffect(() => {
     fetchData()
@@ -33,16 +25,14 @@ const HomePage = () => {
   return (
     <PageLayout>
       <header>
-        <h1>Dashboad Page</h1>
+        <h1>Project Page</h1>
       </header>
-      
-      <section className="dashboad-page">
-        {/* <MonthlyCalendar /> */}
-        <SessionList sessions={sessions} />
-        {/* <Swipable  /> */}
+
+      <section className="project-page">
+        <Tasks projects={state} />
       </section>
     </PageLayout>
-  );
+  )
 };
 
 export default HomePage;
